Rename main module to 'events'
