+++ test/bootstrap.js
@@ -9,8 +9,8 @@ const constants = {
 
 // NOTE: We set the database path to "test", as we don't want tests to run
 // on a clean database each time.
-let main = rewire("../src/main");
-main.__set__("config.database.path", constants.DATABASE.NAME);
+let events = rewire("../src/events");
+events.__set__("config.database.path", constants.DATABASE.NAME);
 
 module.exports = {
   constants
